feat: rename input component
