+++ apps/browser-extension-wallet/src/features/account/components/EditAccount/EditAccountDrawer.tsx
@@ -1,6 +1,6 @@
 import React, { useState } from 'react';
 import { useTranslation } from 'react-i18next';
-import { Box, Flex, Button, Text, Input } from '@lace/ui';
+import { Box, Flex, Button, Text, TextBox } from '@lace/ui';
 import { Drawer, DrawerNavigation } from '@lace/common';
 
 export type Props = {
@@ -47,7 +47,7 @@ export const EditAccountDrawer = ({ name, index, visible, onSave, hide }: Props)
         <Box mb="$64">
           <Text.Body.Normal>{t('account.edit.subtitle')}</Text.Body.Normal>
         </Box>
-        <Input
+        <TextBox
           data-testid="edit-account-name-input"
           containerStyle={{ width: '100%' }}
           label={t('account.edit.input.label')}
