+++ packages/ui/src/design-system/index.ts
@@ -27,4 +27,4 @@ export { PasswordBox } from './password-box';
 export { Metadata } from './metadata';
 export { TextLink } from './text-link';
 export * as ProfileDropdown from './profile-dropdown';
-export { Input } from './input';
+export { TextBox } from './text-box';
