--- packages/ui/src/design-system/input/index.ts
@@ -1,2 +1,0 @@
-export { Input } from './input.component';
-export type { InputProps } from './input.component';
