+++ packages/ui/src/design-system/text-box/index.ts
@@ -1,0 +1,2 @@
+export { TextBox } from './text-box.component';
+export type { TextBoxProps } from './text-box.component';
