+++ packages/ui/src/design-system/text-box/text-box.css.ts
